@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": ">=5.5",
         "ext-mbstring": "*",
         "doctrine/collections": "~1.0",
         "symfony/filesystem": "^2.0.5|^3.0",
@@ -21,12 +21,14 @@
     "require-dev": {
         "ext-zip": "*",
         "guzzle/guzzle": "~3.0",
+        "guzzlehttp/guzzle": "^6.0",
         "phpunit/phpunit": "^4.0|^5.0",
         "symfony/finder": "^2.0.5|^3.0"
     },
     "suggest": {
         "ext-zip": "To use the ZipExtensionAdapter",
-        "guzzle/guzzle": "To use the GuzzleTeleporter"
+        "guzzlehttp/guzzle": "To use the GuzzleTeleporter with Guzzle 6",
+        "guzzle/guzzle": "To use the GuzzleTeleporter with Guzzle 3"
     },
     "autoload": {
         "psr-4": {
@@ -41,7 +43,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "0.2.x-dev"
+            "dev-master": "0.4.x-dev"
         }
     }
 }
