@@ -10,10 +10,10 @@
     ],
     "require": {
         "php": ">=5.3.2",
-        "symfony/console": "~2.2"
+        "symfony/console": "~2.3 || ~3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.1"
+        "phpunit/phpunit": "~4.4"
     },
     "autoload": {
         "psr-4": {
@@ -22,7 +22,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "0.5.x-dev"
+            "dev-master": "0.6.x-dev"
         }
     }
 }
