@@ -1,13 +1,6 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: kris
- * Date: 4/24/15
- * Time: 4:14 PM
- */
 
 namespace Drupal\password_policy\Form;
-
 
 use Drupal\Core\Form\ConfirmFormBase;
 use Drupal\Core\Form\ConfirmFormHelper;
@@ -16,25 +9,36 @@ use Drupal\Core\Url;
 use Drupal\user\SharedTempStoreFactory;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
+/**
+ * Deleting a constraint from a policy within the wizard.
+ */
 class ConstraintDelete extends ConfirmFormBase {
 
   /**
+   * Temp store to maintain state between steps of the wizard.
+   *
    * @var \Drupal\user\SharedTempStoreFactory
    */
   protected $tempstore;
 
   /**
+   * ID of the tempstore to maintain state for the form wizard.
+   *
    * @var string
    */
-  protected $tempstore_id = 'password_policy.password_policy';
+  protected $tempstoreId = 'password_policy.password_policy';
 
   /**
-   * @var string;
+   * The machine name of the form step.
+   *
+   * @var string
    */
-  protected $machine_name;
+  protected $machineName;
 
   /**
-   * @var int;
+   * ID of the constraint.
+   *
+   * @var int
    */
   protected $id;
 
@@ -46,9 +50,12 @@ class ConstraintDelete extends ConfirmFormBase {
   }
 
   /**
+   * Constructor that adds the tempstore from the container for wizard.
+   *
    * @param \Drupal\user\SharedTempStoreFactory $tempstore
+   *   The tempstore of the wizard form.
    */
-  function __construct(SharedTempStoreFactory $tempstore) {
+  public function __construct(SharedTempStoreFactory $tempstore) {
     $this->tempstore = $tempstore;
   }
 
@@ -69,24 +76,28 @@ class ConstraintDelete extends ConfirmFormBase {
    *   An associative array containing the structure of the form.
    * @param \Drupal\Core\Form\FormStateInterface $form_state
    *   The current state of the form.
+   * @param string $machine_name
+   *   The machine name of the policy.
+   * @param int $id
+   *   The ID value of the constraint.
    *
    * @return array
    *   The form structure.
    */
   public function buildForm(array $form, FormStateInterface $form_state, $machine_name = NULL, $id = NULL) {
-    $this->machine_name = $machine_name;
+    $this->machineName = $machine_name;
     $this->id = $id;
 
-    $cached_values = $this->tempstore->get($this->tempstore_id)->get($this->machine_name);
-    $form ['#title'] = $this->getQuestion($id, $cached_values);
+    $cached_values = $this->tempstore->get($this->tempstoreId)->get($this->machineName);
+    $form['#title'] = $this->getQuestion($id, $cached_values);
 
-    $form ['#attributes']['class'][] = 'confirmation';
-    $form ['description'] = array('#markup' => $this->getDescription());
-    $form [$this->getFormName()] = array('#type' => 'hidden', '#value' => 1);
+    $form['#attributes']['class'][] = 'confirmation';
+    $form['description'] = array('#markup' => $this->getDescription());
+    $form[$this->getFormName()] = array('#type' => 'hidden', '#value' => 1);
 
     // By default, render the form using theme_confirm_form().
-    if (!isset($form ['#theme'])) {
-      $form ['#theme'] = 'confirm_form';
+    if (!isset($form['#theme'])) {
+      $form['#theme'] = 'confirm_form';
     }
     $form['actions'] = array('#type' => 'actions');
     $form['actions'] += $this->actions($form, $form_state);
@@ -97,14 +108,14 @@ class ConstraintDelete extends ConfirmFormBase {
    * {@inheritdoc}
    */
   public function submitForm(array &$form, FormStateInterface $form_state) {
-    $cached_values = $this->tempstore->get($this->tempstore_id)->get($this->machine_name);
-    /** @var $policy \Drupal\password_policy\Entity\PasswordPolicy */
+    $cached_values = $this->tempstore->get($this->tempstoreId)->get($this->machineName);
+    /** @var \Drupal\password_policy\Entity\PasswordPolicy $policy */
     $policy = $cached_values['password_policy'];
     $constraints = $policy->getConstraints();
     unset($constraints[$this->id]);
     $policy->set('policy_constraints', $constraints);
-    $this->tempstore->get($this->tempstore_id)->set($this->machine_name, $cached_values);
-    $form_state->setRedirect('entity.password_policy.wizard.edit', ['machine_name' => $this->machine_name, 'step' => 'constraint']);
+    $this->tempstore->get($this->tempstoreId)->set($this->machineName, $cached_values);
+    $form_state->setRedirect('entity.password_policy.wizard.edit', ['machine_name' => $this->machineName, 'step' => 'constraint']);
   }
 
   /**
@@ -126,11 +137,19 @@ class ConstraintDelete extends ConfirmFormBase {
    *   A URL object.
    */
   public function getCancelUrl() {
-    return new Url('entity.password_policy.wizard.edit', ['machine_name' => $this->machine_name, 'step' => 'constraint']);
+    return new Url('entity.password_policy.wizard.edit', ['machine_name' => $this->machineName, 'step' => 'constraint']);
   }
 
   /**
    * Provides the action buttons for submitting this form.
+   *
+   * @param array $form
+   *   An associative array containing the structure of the form.
+   * @param \Drupal\Core\Form\FormStateInterface $form_state
+   *   The current state of the form.
+   *
+   * @return array
+   *   A set of actions associated with this form.
    */
   protected function actions(array $form, FormStateInterface $form_state) {
     return array(
