@@ -1,13 +1,6 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: kris
- * Date: 4/24/15
- * Time: 4:14 PM
- */
 
 namespace Drupal\password_policy\Form;
-
 
 use Drupal\Component\Plugin\PluginManagerInterface;
 use Drupal\Core\Ajax\AjaxResponse;
@@ -15,31 +8,43 @@ use Drupal\Core\Ajax\CloseModalDialogCommand;
 use Drupal\Core\Ajax\RedirectCommand;
 use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Url;
 use Drupal\user\SharedTempStoreFactory;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
+/**
+ * Editing a constraint within the policy wizard form.
+ */
 class ConstraintEdit extends FormBase {
 
 
   /**
+   * Adding a tempstore for the multiple steps of the wizard form.
+   *
    * @var \Drupal\user\SharedTempStoreFactory
    */
   protected $tempstore;
 
   /**
+   * Plugin manager of the password constraints.
+   *
    * @var \Drupal\Component\Plugin\PluginManagerInterface
    */
   protected $manager;
 
   /**
+   * Identifier of the wizard's tempstore.
+   *
    * @var string
    */
-  protected $tempstore_id = 'password_policy.password_policy';
+  protected $tempstoreId = 'password_policy.password_policy';
 
   /**
-   * @var string;
+   * Machine name of the form step.
+   *
+   * @var string
    */
-  protected $machine_name;
+  protected $machineName;
 
   /**
    * {@inheritdoc}
@@ -48,7 +53,15 @@ class ConstraintEdit extends FormBase {
     return new static($container->get('user.shared_tempstore'), $container->get('plugin.manager.password_policy.password_constraint'));
   }
 
-  function __construct(SharedTempStoreFactory $tempstore, PluginManagerInterface $manager) {
+  /**
+   * Overriding the constructor to load in the plugin manager and tempstore.
+   *
+   * @param \Drupal\user\SharedTempStoreFactory $tempstore
+   *   The tempstore of the wizard form.
+   * @param \Drupal\Component\Plugin\PluginManagerInterface $manager
+   *   The plugin manager for the password constraints.
+   */
+  public function __construct(SharedTempStoreFactory $tempstore, PluginManagerInterface $manager) {
     $this->tempstore = $tempstore;
     $this->manager = $manager;
   }
@@ -70,14 +83,18 @@ class ConstraintEdit extends FormBase {
    *   An associative array containing the structure of the form.
    * @param \Drupal\Core\Form\FormStateInterface $form_state
    *   The current state of the form.
+   * @param string $constraint_id
+   *   Plugin ID of the constraint.
+   * @param string $machine_name
+   *   Machine name of this form step.
    *
    * @return array
    *   The form structure.
    */
   public function buildForm(array $form, FormStateInterface $form_state, $constraint_id = NULL, $machine_name = NULL) {
-    $this->machine_name = $machine_name;
-    $cached_values = $this->tempstore->get($this->tempstore_id)->get($this->machine_name);
-    /** @var $policy \Drupal\password_policy\Entity\PasswordPolicy */
+    $this->machineName = $machine_name;
+    $cached_values = $this->tempstore->get($this->tempstoreId)->get($this->machineName);
+    /** @var \Drupal\password_policy\Entity\PasswordPolicy $policy */
     $policy = $cached_values['password_policy'];
     if (is_numeric($constraint_id)) {
       $id = $constraint_id;
@@ -87,25 +104,25 @@ class ConstraintEdit extends FormBase {
     else {
       $instance = $this->manager->createInstance($constraint_id, []);
     }
-    /** @var $instance \Drupal\password_policy\PasswordConstraintInterface */
+    /** @var \Drupal\password_policy\PasswordConstraintInterface $instance */
     $form = $instance->buildConfigurationForm($form, $form_state);
     if (isset($id)) {
       // Conditionally set this form element so that we can update or add.
       $form['id'] = [
         '#type' => 'value',
-        '#value' => $id
+        '#value' => $id,
       ];
     }
     $form['instance'] = [
       '#type' => 'value',
-      '#value' => $instance
+      '#value' => $instance,
     ];
     $form['submit'] = [
       '#type' => 'submit',
       '#value' => $this->t('Save'),
       '#ajax' => [
         'callback' => [$this, 'ajaxSave'],
-      ]
+      ],
     ];
     return $form;
   }
@@ -113,12 +130,20 @@ class ConstraintEdit extends FormBase {
   /**
    * {@inheritdoc}
    */
+  public function validateForm(array &$form, FormStateInterface $form_state) {
+    $instance = $form_state->getValue('instance');
+    $instance->validateConfigurationForm($form, $form_state);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
   public function submitForm(array &$form, FormStateInterface $form_state) {
-    $cached_values = $this->tempstore->get($this->tempstore_id)->get($this->machine_name);
-    /** @var $policy \Drupal\password_policy\Entity\PasswordPolicy */
+    $cached_values = $this->tempstore->get($this->tempstoreId)->get($this->machineName);
+    /** @var \Drupal\password_policy\Entity\PasswordPolicy $policy */
     $policy = $cached_values['password_policy'];
     $constraints = $policy->getConstraints();
-    /** @var $instance \Drupal\password_policy\PasswordConstraintInterface */
+    /** @var \Drupal\password_policy\PasswordConstraintInterface $instance */
     $instance = $form_state->getValue('instance');
     $instance->submitConfigurationForm($form, $form_state);
     if ($form_state->hasValue('id')) {
@@ -128,13 +153,17 @@ class ConstraintEdit extends FormBase {
       $constraints[] = $instance->getConfiguration();
     }
     $policy->set('policy_constraints', $constraints);
-    $this->tempstore->get($this->tempstore_id)->set($this->machine_name, $cached_values);
-    $form_state->setRedirect('entity.password_policy.wizard.edit', ['machine_name' => $this->machine_name, 'step' => 'constraint']);
+    $this->tempstore->get($this->tempstoreId)->set($this->machineName, $cached_values);
+    $form_state->setRedirect('entity.password_policy.wizard.edit', ['machine_name' => $this->machineName, 'step' => 'constraint']);
   }
 
+  /**
+   * {@inheritdoc}
+   */
   public function ajaxSave(array &$form, FormStateInterface $form_state) {
     $response = new AjaxResponse();
-    $response->addCommand(new RedirectCommand($this->url('entity.password_policy.wizard.edit', ['machine_name' => $this->machine_name, 'step' => 'constraint'])));
+    $url = Url::fromRoute('entity.password_policy.wizard.edit', ['machine_name' => $this->machineName, 'step' => 'constraint']);
+    $response->addCommand(new RedirectCommand($url->toString()));
     $response->addCommand(new CloseModalDialogCommand());
     return $response;
   }
