@@ -1,16 +1,15 @@
 <?php
-/**
- * @file
- * Contains \Drupal\password_policy\Form\PasswordPolicyGeneralForm.
- */
 
 namespace Drupal\password_policy\Form;
-
 
 use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
 
+/**
+ * The general settings of the policy not tied to constraints.
+ */
 class PasswordPolicyGeneralForm extends FormBase {
+
   /**
    * {@inheritdoc}
    */
@@ -23,13 +22,13 @@ class PasswordPolicyGeneralForm extends FormBase {
    */
   public function buildForm(array $form, FormStateInterface $form_state) {
     $cached_values = $form_state->getTemporaryValue('wizard');
-    /** @var $policy \Drupal\password_policy\Entity\PasswordPolicy */
+    /** @var \Drupal\password_policy\Entity\PasswordPolicy $policy */
     $policy = $cached_values['password_policy'];
 
     $form['password_reset'] = [
       '#type' => 'textfield',
-      '#title' => t('Password Reset Days'),
-      '#description' => t('User password will reset after the selected number of days'),
+      '#title' => $this->t('Password Reset Days'),
+      '#description' => $this->t('User password will reset after the selected number of days.  0 days indicates that passwords never expire.'),
       '#default_value' => $policy->getPasswordReset(),
     ];
     return $form;
@@ -40,7 +39,7 @@ class PasswordPolicyGeneralForm extends FormBase {
    */
   public function submitForm(array &$form, FormStateInterface $form_state) {
     $cached_values = $form_state->getTemporaryValue('wizard');
-    /** @var $policy \Drupal\password_policy\Entity\PasswordPolicy */
+    /** @var \Drupal\password_policy\Entity\PasswordPolicy $policy */
     $policy = $cached_values['password_policy'];
     $policy->set('password_reset', $form_state->getValue('password_reset'));
     $form_state->setTemporaryValue('wizard', $cached_values);
