@@ -1,8 +1,10 @@
 <?php
+
 namespace Drupal\password_policy;
 
 /**
- * Class PasswordPolicyValidation
+ * A construct to organize validation of a password policy.
+ *
  * @package Drupal\password_policy
  */
 class PasswordPolicyValidation {
@@ -10,27 +12,32 @@ class PasswordPolicyValidation {
   protected $error = NULL;
   protected $valid = TRUE;
 
-  /*
-   * Set error message and mark as invalid
+  /**
+   * Set error message and mark as invalid.
    */
   public function setErrorMessage($error) {
     $this->valid = FALSE;
     $this->error = $error;
   }
 
-  /*
-   * Output error message
+  /**
+   * Output error message.
+   *
    * @return string
+   *   A message representing the error message of the policy's constraints.
    */
   public function getErrorMessage() {
     return $this->error;
   }
 
-  /*
-   * Output validation state
-   * @return boolean
+  /**
+   * Output validation state.
+   *
+   * @return bool
+   *   Whether or not the policy has an error or not.
    */
   public function isValid() {
     return $this->valid;
   }
+
 }
